update linear merging
